New Text module to consolidate string functions.
