--- test/KeymapSpec.hs
@@ -1,20 +1,0 @@
-module KeymapSpec (tests) where
-
-import Test.Tasty
-import Test.Tasty.HUnit
-
-import Keymap (dropLastUTF8)
-import UI (u)
-
-tests :: TestTree
-tests = testGroup "Keymap"
-    [ testGroup "dropLastUTF8"
-        [ testCase "ASCII"    $ dropLastUTF8 "abc"          @?= "ab"
-        , testCase "empty"    $ dropLastUTF8 ""             @?= ""
-        , testCase "French"   $ dropLastUTF8 (u"été")       @?= u"ét"
-        , testCase "CJK"      $ dropLastUTF8 (u"中國")      @?= u"中"
-        , testCase "Emoji"    $ dropLastUTF8 (u"Yes👍")     @?= u"Yes"
-        , testCase "invalid"  $ dropLastUTF8 "\x80"         @?= ""
-        ]
-    ]
-
