+++ test/Main.hs
@@ -2,24 +2,20 @@ module Main (main) where
 
 import Test.Tasty
 
-import BaseSpec qualified
 import ConfigSpec qualified
 import CoreSpec qualified
 import DecoderSpec qualified
-import KeymapSpec qualified
 import PlaylistSpec qualified
 import StyleSpec qualified
-import WidthSpec qualified
+import TextSpec qualified
 
 main :: IO ()
 main = defaultMain $ testGroup "hmp3-ng"
-    [ BaseSpec.tests
-    , ConfigSpec.tests
+    [ ConfigSpec.tests
     , CoreSpec.tests
     , DecoderSpec.tests
-    , KeymapSpec.tests
     , StyleSpec.tests
     , PlaylistSpec.tests
-    , WidthSpec.tests
+    , TextSpec.tests
     ]
